+++ packages/web/components/Player/Section/PlayerGallery.tsx
@@ -15,17 +15,15 @@ import {
   useDisclosure,
   ViewAllButton,
 } from '@metafam/ds';
-import { MetaLink as Link } from 'components/Link';
 import { ProfileSection } from 'components/Section/ProfileSection';
 import { Player } from 'graphql/autogen/types';
-import { useOpenSeaCollectibles } from 'lib/hooks/opensea';
 import { BoxTypes } from 'utils/boxTypes';
-import { Collectible } from 'utils/openseaHelpers';
+import { useNFTCollectibles } from 'lib/hooks/alchemy';
 
-const GalleryItem: React.FC<{ nft: Collectible }> = ({ nft }) => (
-  <Link href={nft.openseaLink} isExternal display="flex">
+const GalleryItem: React.FC<{ nft: any }> = ({ nft }) => (
+  <Box display="flex">
     <Box
-      bgImage={`url(${nft.imageURL})`}
+      bgImage={`url(${nft.image.cachedUrl})`}
       backgroundSize="contain"
       backgroundRepeat="no-repeat"
       backgroundPosition="center"
@@ -49,18 +47,18 @@ const GalleryItem: React.FC<{ nft: Collectible }> = ({ nft }) => (
             fontVariant: 'small-caps',
           }}
         >
-          {nft.title}
+          {nft?.collection?.name || nft?.contract?.name || 'Unknown'}
         </Heading>
-        <Text fontSize="sm">{nft.priceString}</Text>
+        <Text fontSize="sm">Floor Price: {nft.contract.openSeaMetadata.floorPrice || '0' } ETH</Text>
       </Flex>
     </Tooltip>
-  </Link>
+  </Box>
 );
 
 type GalleryModalProps = {
   isOpen: boolean;
   onClose: () => void;
-  nfts: Array<Collectible>;
+  nfts: Array<any>;
 };
 
 const GalleryModal: React.FC<GalleryModalProps> = ({
@@ -97,15 +95,8 @@ export const PlayerGallery: React.FC<Props> = ({
   editing,
 }) => {
   const { isOpen, onOpen, onClose } = useDisclosure();
-  const {
-    favorites,
-    data: nfts,
-    loading,
-    error,
-  } = useOpenSeaCollectibles({
-    player,
-  });
-
+  const {loading, error, data: nfts} = useNFTCollectibles({ player }); 
+  console.log(nfts)
   return (
     <ProfileSection
       title="NFT Gallery"
@@ -133,14 +124,14 @@ export const PlayerGallery: React.FC<Props> = ({
         return (
           <>
             <SimpleGrid columns={1} gap={4} px={2}>
-              {favorites?.map((nft) => (
+              {nfts[0]?.nfts?.ownedNfts?.map((nft: any) => (
                 <GalleryItem {...{ nft }} key={nft.tokenId} />
               ))}
             </SimpleGrid>
-            {nfts.length > 3 && (
+            {nfts[0]?.nfts?.ownedNfts.length > 3 && (
               <Box textAlign="end">
                 <GalleryModal {...{ isOpen, onClose, nfts }} />
-                <ViewAllButton onClick={onOpen} size={nfts.length} />
+                <ViewAllButton onClick={onOpen} size={nfts[0]?.nfts?.ownedNfts.length} />
               </Box>
             )}
           </>
