+++ packages/web/package.json
@@ -35,6 +35,7 @@
     "@walletconnect/web3-provider": "1.7.3",
     "@web3-storage/access": "^18.1.1",
     "@web3-storage/w3up-client": "^12.0.0",
+    "alchemy-sdk": "^3.1.2",
     "chakra-ui-markdown-renderer": "^4.1.0",
     "copy-to-clipboard": "3.3.1",
     "deep-equal": "2.0.5",
@@ -52,7 +53,6 @@
     "next": "^13.3.0",
     "next-plausible": "^3.7.2",
     "next-urql": "^5.0.2",
-    "opensea-js": "^4.0.12",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-dropzone": "^14.2.2",
